+++ src/Propulsion.Cosmos/EquinoxCosmosParser.fs
@@ -14,6 +14,26 @@ open Propulsion.Streams
 /// Maps fields in an Event within an Equinox.Cosmos V1+ Event (in a Batch or Tip) to the interface defined by Propulsion.Streams
 /// <remarks>NOTE No attempt is made to filter out Tip (`id=-1`) batches from the ChangeFeed; Equinox versions >= 3, Tip batches can bear events.</remarks>
 [<RequireQualifiedAccess>]
+#if !COSMOSV2 && !COSMOSV3
+module EquinoxSystemTextJsonParser =
+
+    type System.Text.Json.JsonDocument with
+        member document.Cast<'T>() =
+            System.Text.Json.JsonSerializer.Deserialize<'T>(document.RootElement)
+    type Batch with
+        member _.MapData x =
+            System.Text.Json.JsonSerializer.SerializeToUtf8Bytes x
+    let timestamp (doc : System.Text.Json.JsonDocument) =
+        let unixEpoch = System.DateTime.UnixEpoch
+        let ts = let r = doc.RootElement in r.GetProperty("_ts")
+        unixEpoch.AddSeconds(ts.GetDouble())
+
+    /// Sanity check to determine whether the Document represents an `Equinox.Cosmos` >= 1.0 based batch
+    let isEquinoxBatch (d : System.Text.Json.JsonDocument) =
+        let r = d.RootElement
+        let hasProp (id : string) = match r.TryGetProperty id with true, _ -> true | _ -> false
+        hasProp "p" && hasProp "i" && hasProp "n" && hasProp "e"
+#else
 #if COSMOSV2
 module EquinoxCosmosParser =
 
@@ -22,6 +42,8 @@ module EquinoxCosmosParser =
             let tmp = Document()
             tmp.SetPropertyValue("content", document)
             tmp.GetPropertyValue<'T>("content")
+    type Batch with
+        member _.MapData x = x
 
     /// Sanity check to determine whether the Document represents an `Equinox.Cosmos` >= 1.0 based batch
     let isEquinoxBatch (d : Document) =
@@ -33,6 +55,8 @@ module EquinoxNewtonsoftParser =
     type Newtonsoft.Json.Linq.JObject with
         member document.Cast<'T>() =
             document.ToObject<'T>()
+    type Batch with
+        member _.MapData x = x
 
     let timestamp (doc : Newtonsoft.Json.Linq.JObject) =
         let unixEpoch = System.DateTime.UnixEpoch
@@ -42,11 +66,12 @@ module EquinoxNewtonsoftParser =
     let isEquinoxBatch (d : Newtonsoft.Json.Linq.JObject) =
         d.ContainsKey "p" && d.ContainsKey "i" && d.ContainsKey "n" && d.ContainsKey "e"
 #endif
+#endif
 
     /// Enumerates the events represented within a batch
     let enumEquinoxCosmosEvents (batch : Batch) : StreamEvent<byte[]> seq =
         let streamName = FsCodec.StreamName.parse batch.p // we expect all Equinox data to adhere to "{category}-{aggregateId}" form (or we'll throw)
-        batch.e |> Seq.mapi (fun offset x -> { stream = streamName; event = FsCodec.Core.TimelineEvent.Create(batch.i+int64 offset, x.c, x.d, x.m, timestamp=x.t) })
+        batch.e |> Seq.mapi (fun offset x -> { stream = streamName; event = FsCodec.Core.TimelineEvent.Create(batch.i+int64 offset, x.c, batch.MapData x.d, batch.MapData x.m, timestamp=x.t) })
 
     /// Collects all events with a Document [typically obtained via the CosmosDb ChangeFeed] that potentially represents an Equinox.Cosmos event-batch
     let enumStreamEvents d : StreamEvent<byte[]> seq =