+++ src/Propulsion.EventStore/EventStoreSink.fs
@@ -1,6 +1,13 @@
-﻿namespace Propulsion.EventStore
+﻿#if EVENTSTORE_LEGACY
+namespace Propulsion.EventStore
 
 open Equinox.EventStore
+#else
+namespace Propulsion.EventStoreDb
+
+open Equinox.EventStoreDb
+#endif
+
 open Propulsion
 open Propulsion.Streams
 open Propulsion.Streams.Internal // Helpers
@@ -10,15 +17,13 @@ open System
 open System.Threading
 
 [<AutoOpen>]
-module private Impl =
+module private Impl2 =
     let inline mb x = float x / 1024. / 1024.
 
 module Internal =
 
     [<AutoOpen>]
     module Writer =
-        type EventStoreContext() = class end
-
         type [<NoComparison; NoEquality>] Result =
             | Ok of updatedPos : int64
             | Duplicate of updatedPos : int64
@@ -27,26 +32,34 @@ module Internal =
 
         let logTo (log : ILogger) (res : FsCodec.StreamName * Choice<EventMetrics * Result, EventMetrics * exn>) =
             match res with
-            | stream, (Choice1Of2 (_, Ok pos)) ->
+            | stream, Choice1Of2 (_, Ok pos) ->
                 log.Information("Wrote     {stream} up to {pos}", stream, pos)
-            | stream, (Choice1Of2 (_, Duplicate updatedPos)) ->
+            | stream, Choice1Of2 (_, Duplicate updatedPos) ->
                 log.Information("Ignored   {stream} (synced up to {pos})", stream, updatedPos)
-            | stream, (Choice1Of2 (_, PartialDuplicate overage)) ->
+            | stream, Choice1Of2 (_, PartialDuplicate overage) ->
                 log.Information("Requeuing {stream} {pos} ({count} events)", stream, overage.index, overage.events.Length)
-            | stream, (Choice1Of2 (_, PrefixMissing (batch, pos))) ->
+            | stream, Choice1Of2 (_, PrefixMissing (batch, pos)) ->
                 log.Information("Waiting   {stream} missing {gap} events ({count} events @ {pos})", stream, batch.index - pos, batch.events.Length, batch.index)
-            | stream, (Choice2Of2 (_, exn)) ->
+            | stream, Choice2Of2 (_, exn) ->
                 log.Warning(exn,"Writing   {stream} failed, retrying", stream)
 
-        let write (log : ILogger) (context : Equinox.EventStore.EventStoreContext) stream span = async {
+        let write (log : ILogger) (context : EventStoreContext) stream span = async {
             log.Debug("Writing {s}@{i}x{n}", stream, span.index, span.events.Length)
             let! res = context.Sync(log, stream, span.index - 1L, span.events |> Array.map (fun x -> x :> _))
             let ress =
                 match res with
                 | GatewaySyncResult.Written (Token.Unpack pos') ->
+#if EVENTSTORE_LEGACY
                     Ok (pos'.pos.streamVersion + 1L)
+#else
+                    Ok (pos'.streamVersion + 1L)
+#endif
                 | GatewaySyncResult.ConflictUnknown (Token.Unpack pos) ->
+#if EVENTSTORE_LEGACY
                     match pos.pos.streamVersion + 1L with
+#else
+                    match pos.streamVersion + 1L with
+#endif
                     | actual when actual < span.index -> PrefixMissing (span, actual)
                     | actual when actual >= span.index + span.events.LongLength -> Duplicate actual
                     | actual -> PartialDuplicate { index = actual; events = span.events |> Array.skip (actual - span.index |> int) }
@@ -57,31 +70,30 @@ module Internal =
 
         let classify e =
             match box e with
-            | :? System.TimeoutException -> ResultKind.TimedOut
+            | :? TimeoutException -> ResultKind.TimedOut
             | _ -> ResultKind.Other
 
     type Stats(log : ILogger, statsInterval, stateInterval) =
         inherit Scheduling.Stats<EventMetrics * Writer.Result, EventMetrics * exn>(log, statsInterval, stateInterval)
-        let okStreams, resultOk, resultDup, resultPartialDup, resultPrefix, resultExnOther = HashSet(), ref 0, ref 0, ref 0, ref 0, ref 0
-        let badCats, failStreams, timedOut = CatStats(), HashSet(), ref 0
-        let toStreams, oStreams = HashSet(), HashSet()
+        let mutable okStreams, badCats, failStreams, toStreams, oStreams = HashSet(), CatStats(), HashSet(), HashSet(), HashSet()
+        let mutable resultOk, resultDup, resultPartialDup, resultPrefix, resultExnOther, timedOut = 0, 0, 0, 0, 0, 0
         let mutable okEvents, okBytes, exnEvents, exnBytes = 0, 0L, 0, 0L
 
         override _.DumpStats() =
-            let results = !resultOk + !resultDup + !resultPartialDup + !resultPrefix
+            let results = resultOk + resultDup + resultPartialDup + resultPrefix
             log.Information("Completed {mb:n0}MB {completed:n0}r {streams:n0}s {events:n0}e ({ok:n0} ok {dup:n0} redundant {partial:n0} partial {prefix:n0} waiting)",
-                mb okBytes, results, okStreams.Count, okEvents, !resultOk, !resultDup, !resultPartialDup, !resultPrefix)
-            okStreams.Clear(); resultOk := 0; resultDup := 0; resultPartialDup := 0; resultPrefix := 0; okEvents <- 0; okBytes <- 0L
-            if !timedOut <> 0 || badCats.Any then
-                let fails = !timedOut + !resultExnOther
+                mb okBytes, results, okStreams.Count, okEvents, resultOk, resultDup, resultPartialDup, resultPrefix)
+            okStreams.Clear(); resultOk <- 0; resultDup <- 0; resultPartialDup <- 0; resultPrefix <- 0; okEvents <- 0; okBytes <- 0L
+            if timedOut <> 0 || badCats.Any then
+                let fails = timedOut + resultExnOther
                 log.Warning("Exceptions {mb:n0}MB {fails:n0}r {streams:n0}s {events:n0}e Timed out {toCount:n0}r {toStreams:n0}s",
-                    mb exnBytes, fails, failStreams.Count, exnEvents, !timedOut, toStreams.Count)
-                timedOut := 0; resultExnOther := 0; failStreams.Clear(); toStreams.Clear(); exnBytes <- 0L; exnEvents <- 0
+                    mb exnBytes, fails, failStreams.Count, exnEvents, timedOut, toStreams.Count)
+                timedOut <- 0; resultExnOther <- 0; failStreams.Clear(); toStreams.Clear(); exnBytes <- 0L; exnEvents <- 0
             if badCats.Any then
                 log.Warning(" Affected cats {@badCats} Other {other:n0}r {@oStreams}",
-                    badCats.StatsDescending |> Seq.truncate 50, !resultExnOther, oStreams |> Seq.truncate 100)
-                badCats.Clear(); resultExnOther := 0; oStreams.Clear()
-            Equinox.EventStore.Log.InternalMetrics.dump log
+                    badCats.StatsDescending |> Seq.truncate 50, resultExnOther, oStreams |> Seq.truncate 100)
+                badCats.Clear(); resultExnOther <- 0; oStreams.Clear()
+            Log.InternalMetrics.dump log
 
         override this.Handle message =
             let inline adds x (set : HashSet<_>) = set.Add x |> ignore
@@ -94,18 +106,18 @@ module Internal =
                 okEvents <- okEvents + es
                 okBytes <- okBytes + int64 bs
                 match res with
-                | Writer.Result.Ok _ -> incr resultOk
-                | Writer.Result.Duplicate _ -> incr resultDup
-                | Writer.Result.PartialDuplicate _ -> incr resultPartialDup
-                | Writer.Result.PrefixMissing _ -> incr resultPrefix
+                | Writer.Result.Ok _ -> resultOk <- resultOk + 1
+                | Writer.Result.Duplicate _ -> resultDup <- resultDup + 1
+                | Writer.Result.PartialDuplicate _ -> resultPartialDup <- resultPartialDup + 1
+                | Writer.Result.PrefixMissing _ -> resultPrefix <- resultPrefix + 1
                 this.HandleOk res
             | Scheduling.InternalMessage.Result (_duration, stream, _progressed, Choice2Of2 ((es, bs), exn)) ->
                 adds stream failStreams
                 exnEvents <- exnEvents + es
                 exnBytes <- exnBytes + int64 bs
                 match Writer.classify exn with
-                | ResultKind.TimedOut -> adds stream toStreams; incr timedOut
-                | ResultKind.Other -> bads stream oStreams; incr resultExnOther
+                | ResultKind.TimedOut -> adds stream toStreams; timedOut <- timedOut + 1
+                | ResultKind.Other -> bads stream oStreams; resultExnOther <- resultExnOther + 1
                 this.HandleExn(log.ForContext("stream", stream).ForContext("events", es), exn)
         abstract member HandleOk : Result -> unit
         default _.HandleOk _ : unit = ()
@@ -120,7 +132,7 @@ module Internal =
 
             let attemptWrite (stream, span) = async {
                 let index = Interlocked.Increment(&robin) % connections.Length
-                let selectedConnection = connections.[index]
+                let selectedConnection = connections[index]
                 let maxEvents, maxBytes = 65536, 4 * 1024 * 1024 - (*fudge*)4096
                 let met, span' = Buffering.StreamSpan.slice (maxEvents, maxBytes) span
                 try let! res = Writer.write storeLog selectedConnection (FsCodec.StreamName.toString stream) span'