+++ src/Propulsion.Feed/FeedSource.fs
@@ -32,15 +32,36 @@ type FeedSourceBase internal
     member internal _.Pump
         (   readTranches : unit -> Async<TrancheId[]>,
             /// Responsible for managing retries and back offs; yielding an exception will result in abend of the read loop
-            crawl : TrancheId -> bool * Position -> AsyncSeq<Internal.Batch<byte[]>>) = async {
+            crawl : TrancheId -> bool * Position -> AsyncSeq<Batch<byte[]>>) = async {
         try let! tranches = readTranches ()
             log.Information("Starting {tranches} tranche readers...", tranches.Length)
-            let crawl trancheId (wasLast, pos) = asyncSeq {
-                yield! crawl trancheId (wasLast, pos) }
             return! Async.Parallel(tranches |> Seq.mapi (pumpPartition crawl)) |> Async.Ignore<unit[]>
         with e ->
             log.Warning(e, "Exception encountered while running source, exiting loop")
             return! Async.Raise e }
+
+/// Drives reading and checkpointing from a singular source that contains data from multiple streams,
+/// such as the SqlStreamStore and EventStoreDB $all feeds
+type AllFeedSource
+    (   log : Serilog.ILogger, statsInterval : TimeSpan,
+        sourceId, tailSleepInterval : TimeSpan,
+        checkpoints : IFeedCheckpointStore, defaultCheckpointEventInterval : TimeSpan,
+        readBatch : TrancheId * Position -> Async<Batch<_>>,
+        sink : ProjectorPipeline<Ingestion.Ingester<seq<StreamEvent<byte[]>>, Submission.SubmissionBatch<int,StreamEvent<byte[]>>>>) =
+    inherit FeedSourceBase(log, statsInterval, sourceId, checkpoints, defaultCheckpointEventInterval, sink)
+
+    let crawl trancheId (wasLast, pos) = asyncSeq {
+        if wasLast then
+            do! Async.Sleep tailSleepInterval
+        try let! batch = readBatch (trancheId, pos)
+            yield batch
+        with e ->
+            log.Warning(e, "Read failure")
+            do! Async.Sleep tailSleepInterval }
+
+    member _.Pump(consumerGroupName) =
+        let readTranches () = async { return [| TrancheId.parse consumerGroupName |] }
+        base.Pump(readTranches, crawl)
 
 namespace Propulsion.Feed
 
@@ -71,8 +92,7 @@ type FeedSource
                 do! Async.Sleep tailSleepInterval
             let! page = readPage (trancheId, pos)
             let items' = page.items |> Array.map (fun x -> { stream = streamName; event = x })
-            let slice : Internal.Batch<_> = { items = items'; checkpoint = page.checkpoint; isTail = page.isTail }
-            yield slice
+            yield ({ items = items'; checkpoint = page.checkpoint; isTail = page.isTail } : Internal.Batch<_>)
         }
 
     /// Drives the continual loop of reading and checkpointing each tranche until a fault occurs. <br/>