+++ src/Propulsion.SqlStreamStore/SqlStreamStoreSource.fs
@@ -2,32 +2,32 @@ namespace Propulsion.SqlStreamStore
 
 open System
 
+type StreamEvent = Propulsion.Streams.StreamEvent<byte[]>
+
 module private Impl =
 
     open Propulsion.Infrastructure // AwaitTaskCorrect
-    open Propulsion.Feed
 
-    let toTimelineEvent (dataJson : string) (msg: SqlStreamStore.Streams.StreamMessage) =
+    let toStreamEvent (dataJson : string) (msg: SqlStreamStore.Streams.StreamMessage) : StreamEvent =
         let inline len0ToNull (x : _[]) = match x with null -> null | x when x.Length = 0 -> null | x -> x
-        FsCodec.Core.TimelineEvent.Create
-            (   int64 msg.StreamVersion,
-                msg.Type,
-                (match dataJson with null -> null | x -> x |> System.Text.Encoding.UTF8.GetBytes |> len0ToNull),
-                msg.JsonMetadata |> System.Text.Encoding.UTF8.GetBytes |> len0ToNull,
-                msg.MessageId,
-                timestamp = DateTimeOffset(msg.CreatedUtc))
-    let readWithDataAsTimelineEvent (msg : SqlStreamStore.Streams.StreamMessage) = async {
+        let e = FsCodec.Core.TimelineEvent.Create
+                  ( int64 msg.StreamVersion,
+                    msg.Type,
+                    (match dataJson with null -> null | x -> x |> System.Text.Encoding.UTF8.GetBytes |> len0ToNull),
+                    msg.JsonMetadata |> System.Text.Encoding.UTF8.GetBytes |> len0ToNull,
+                    msg.MessageId,
+                    timestamp = DateTimeOffset(msg.CreatedUtc))
+        { stream = Propulsion.Streams.StreamName.internalParseSafe msg.StreamId; event = e }
+    let readWithDataAsStreamEvent (msg : SqlStreamStore.Streams.StreamMessage) = async {
         let! json = msg.GetJsonData() |> Async.AwaitTaskCorrect
-        return toTimelineEvent json msg }
-    let readPage excludeBodies maxBatchSize (store : SqlStreamStore.IStreamStore) (_tranche, pos) : Async<Page<_>> = async {
+        return toStreamEvent json msg }
+    let readPage excludeBodies maxBatchSize (store : SqlStreamStore.IStreamStore) (_tranche, pos) : Async<Propulsion.Feed.Internal.Batch<_>> = async {
         let! ct = Async.CancellationToken
-        let! page = store.ReadAllForwards(Position.toInt64 pos, maxBatchSize, not excludeBodies, ct) |> Async.AwaitTaskCorrect
+        let! page = store.ReadAllForwards(Propulsion.Feed.Position.toInt64 pos, maxBatchSize, not excludeBodies, ct) |> Async.AwaitTaskCorrect
         let! items =
-            if excludeBodies then async { return page.Messages |> Array.map (toTimelineEvent null) }
-            else page.Messages |> Seq.map readWithDataAsTimelineEvent |> Async.Sequential
-        return { checkpoint = Position.parse page.NextPosition; items = items; isTail = page.IsEnd } }
-
-type StreamEvent = Propulsion.Streams.StreamEvent<byte[]>
+            if excludeBodies then async { return page.Messages |> Array.map (toStreamEvent null) }
+            else page.Messages |> Seq.map readWithDataAsStreamEvent |> Async.Sequential
+        return { checkpoint = Propulsion.Feed.Position.parse page.NextPosition; items = items; isTail = page.IsEnd } }
 
 type SqlStreamStoreSource
     (   log : Serilog.ILogger, statsInterval : TimeSpan,
@@ -37,10 +37,6 @@ type SqlStreamStoreSource
         sink : Propulsion.ProjectorPipeline<Propulsion.Ingestion.Ingester<seq<StreamEvent>, Propulsion.Submission.SubmissionBatch<int, StreamEvent>>>,
         /// If the Handler does not require the bodies of the events, we can avoid querying the bodies from SqlStreamStore in the first instance
         ?excludeBodies) =
-    inherit Propulsion.Feed.FeedSource(log, statsInterval, sourceId, tailSleepInterval,
-                                       checkpoints, defaultCheckpointEventInterval,
-                                       Impl.readPage (excludeBodies = Some true) maxBatchSize store, sink)
-
-    member _.Pump(consumerGroupName) =
-        let readTranches () = async { return [| Propulsion.Feed.TrancheId.parse consumerGroupName |] }
-        base.Pump(readTranches)
+    inherit Propulsion.Feed.Internal.AllFeedSource(log, statsInterval, sourceId, tailSleepInterval,
+                                                   checkpoints, defaultCheckpointEventInterval,
+                                                   Impl.readPage (excludeBodies = Some true) maxBatchSize store, sink)